create class for convert_switches_cim
